:recycle: refactor: ERD 반영하여 domain 수정

+++ src/main/java/com/gdgoc/bondly/domain/Friendship.java
@@ -15,7 +15,7 @@ import lombok.Setter;
 public class Friendship {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer friendshipId;
+    private Long friendshipId;
     @ManyToOne
     @JoinColumn(name = "sender_id", nullable = false)
     private User sender;
